- Extracted admin role check to method - User methods trait
