+++ app/Http/Requests/Backend/Auth/Role/ManageRoleRequest.php
@@ -16,7 +16,7 @@ class ManageRoleRequest extends FormRequest
      */
     public function authorize()
     {
-        return auth()->user()->hasRole('administrator');
+        return $this->user()->isAdmin();
     }
 
     /**
