+++ app/Http/Requests/Backend/Auth/Role/StoreRoleRequest.php
@@ -16,7 +16,7 @@ class StoreRoleRequest extends FormRequest
      */
     public function authorize()
     {
-        return auth()->user()->hasRole('administrator');
+        return $this->user()->isAdmin();
     }
 
     /**
