+++ app/Http/Requests/Backend/Auth/User/ManageUserRequest.php
@@ -16,7 +16,7 @@ class ManageUserRequest extends FormRequest
      */
     public function authorize()
     {
-        return auth()->user()->hasRole('administrator');
+        return $this->user()->isAdmin();
     }
 
     /**
