+++ app/Http/Requests/Backend/Auth/User/StoreUserRequest.php
@@ -17,7 +17,7 @@ class StoreUserRequest extends FormRequest
      */
     public function authorize()
     {
-        return auth()->user()->hasRole('administrator');
+        return $this->user()->isAdmin();
     }
 
     /**
