+++ app/Http/Requests/Backend/Auth/User/UpdateUserPasswordRequest.php
@@ -16,7 +16,7 @@ class UpdateUserPasswordRequest extends FormRequest
      */
     public function authorize()
     {
-        return auth()->user()->hasRole('administrator');
+        return $this->user()->isAdmin();
     }
 
     /**
