+++ app/Http/Requests/Backend/Auth/User/UpdateUserRequest.php
@@ -16,7 +16,7 @@ class UpdateUserRequest extends FormRequest
      */
     public function authorize()
     {
-        return auth()->user()->hasRole('administrator');
+        return $this->user()->isAdmin();
     }
 
     /**
