+++ app/Models/Auth/Traits/Attribute/UserAttribute.php
@@ -10,6 +10,7 @@ namespace App\Models\Auth\Traits\Attribute;
  */
 trait UserAttribute
 {
+
     /**
      * @return string
      */
@@ -72,46 +73,6 @@ trait UserAttribute
     }
 
     /**
-     * @return mixed
-     */
-    public function canChangeEmail()
-    {
-        return config('access.users.change_email');
-    }
-
-    /**
-     * @return bool
-     */
-    public function canChangePassword()
-    {
-        return ! app('session')->has(config('access.socialite_session_name'));
-    }
-
-    /**
-     * @return bool
-     */
-    public function isActive()
-    {
-        return $this->active == 1;
-    }
-
-    /**
-     * @return bool
-     */
-    public function isConfirmed()
-    {
-        return $this->confirmed == 1;
-    }
-
-    /**
-     * @return bool
-     */
-    public function isPending()
-    {
-        return config('access.users.requires_approval') && $this->confirmed == 0;
-    }
-
-    /**
      * @return string
      */
     public function getFullNameAttribute()
@@ -135,36 +96,6 @@ trait UserAttribute
     public function getPictureAttribute()
     {
         return $this->getPicture();
-    }
-
-    /**
-     * @param bool $size
-     *
-     * @return mixed
-     */
-    public function getPicture($size = false)
-    {
-        if (! $size) {
-            $size = config('gravatar.default.size');
-        }
-
-        return gravatar()->get($this->email, ['size' => $size]);
-    }
-
-    /**
-     * @param $provider
-     *
-     * @return bool
-     */
-    public function hasProvider($provider)
-    {
-        foreach ($this->providers as $p) {
-            if ($p->provider == $provider) {
-                return true;
-            }
-        }
-
-        return false;
     }
 
     /**
