+++ app/Models/Auth/Traits/Method/UserMethod.php
@@ -1,0 +1,88 @@
+<?php
+
+namespace App\Models\Auth\Traits\Method;
+
+/**
+ * Trait UserFunction
+ *
+ * @package App\Models\Auth\TraUserFunction
+ */
+trait UserMethod
+{
+	/**
+	 * @return mixed
+	 */
+	public function canChangeEmail()
+	{
+		return config('access.users.change_email');
+	}
+
+	/**
+	 * @return bool
+	 */
+	public function canChangePassword()
+	{
+		return ! app('session')->has(config('access.socialite_session_name'));
+	}
+
+	/**
+	 * @param bool $size
+	 *
+	 * @return mixed
+	 */
+	public function getPicture($size = false)
+	{
+		if (! $size) {
+			$size = config('gravatar.default.size');
+		}
+
+		return gravatar()->get($this->email, ['size' => $size]);
+	}
+
+	/**
+	 * @param $provider
+	 *
+	 * @return bool
+	 */
+	public function hasProvider($provider)
+	{
+		foreach ($this->providers as $p) {
+			if ($p->provider == $provider) {
+				return true;
+			}
+		}
+
+		return false;
+	}
+
+	/**
+	 * @return mixed
+	 */
+	public function isAdmin() {
+		return $this->hasRole(config('access.users.admin_role'));
+	}
+
+	/**
+	 * @return bool
+	 */
+	public function isActive()
+	{
+		return $this->active == 1;
+	}
+
+	/**
+	 * @return bool
+	 */
+	public function isConfirmed()
+	{
+		return $this->confirmed == 1;
+	}
+
+	/**
+	 * @return bool
+	 */
+	public function isPending()
+	{
+		return config('access.users.requires_approval') && $this->confirmed == 0;
+	}
+}
