+++ app/Models/Auth/User.php
@@ -2,6 +2,7 @@
 
 namespace App\Models\Auth;
 
+use App\Models\Auth\Traits\Method\UserMethod;
 use App\Models\Traits\Uuid;
 use Spatie\Permission\Traits\HasRoles;
 use Illuminate\Notifications\Notifiable;
@@ -22,6 +23,7 @@ class User extends Authenticatable
         SendUserPasswordReset,
         SoftDeletes,
         UserAttribute,
+		UserMethod,
         UserRelationship,
         UserScope,
         Uuid;
