+++ config/access.php
@@ -38,6 +38,11 @@ return [
          */
         'change_email' => env('CHANGE_EMAIL', false),
 
+		/*
+		 * The name of the super administrator role
+		 */
+		'admin_role' => 'administrator',
+
         /*
          * The default role all new registered users get added to
          */
