+++ database/seeds/Auth/PermissionRoleTableSeeder.php
@@ -21,7 +21,7 @@ class PermissionRoleTableSeeder extends Seeder
         $this->disableForeignKeys();
 
         // Create Roles
-        $admin = Role::create(['name' => 'administrator']);
+        $admin = Role::create(['name' => config('access.users.admin_role')]);
         $executive = Role::create(['name' => 'executive']);
         $user = Role::create(['name' => 'user']);
